NomNom-Backend/pyproject.toml
[project]
name = "nomnom-backend"
version = "0.1.0"
description = "API server for NomNom — AI-powered food tracker with a roasting cat"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "anthropic>=0.40.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
    "pyjwt>=2.10.0",
    "bcrypt>=4.2.0",
    "apscheduler>=3.10.0",
    "email-validator>=2.0.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.0",
    "pgvector>=0.3.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.8.0",
    "aiosqlite>=0.20.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/models/*.py" = ["F821"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
